mcp-server/requirements.txt
mcp
qdrant-client==1.12.2
mysql-connector-python
requests
